CMake configured for Lab 15
