+++ Practice/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.0)
+
+project(Semester_1)
+
+add_subdirectory(Lab_15)
