+++ Practice/Lab_15/output.bmp
